--- packages/developer-portal/src/components/api-docs/index.ts
@@ -1,2 +1,0 @@
-import APIDocs from './api-docs'
-export default APIDocs
