+++ packages/developer-portal/src/components/api/__tests__/index.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import { Api } from '../index'
+import { render } from '../../../tests/react-testing'
+
+describe('Api', () => {
+  it('should match a snapshot', () => {
+    expect(render(<Api />)).toMatchSnapshot()
+  })
+})
