+++ packages/developer-portal/src/components/graphql/index.tsx
@@ -1,27 +1,9 @@
-import {
-  SmallText,
-  Button,
-  elHFull,
-  elMb5,
-  elMb9,
-  FlexContainer,
-  Icon,
-  Loader,
-  PageContainer,
-  SecondaryNav,
-  SecondaryNavContainer,
-  SecondaryNavItem,
-  Subtitle,
-  Title,
-} from '@reapit/elements'
+import { Loader, Title } from '@reapit/elements'
 import React, { Dispatch, FC, MutableRefObject, SetStateAction, useEffect, useRef, useState } from 'react'
-import { useHistory, useLocation } from 'react-router'
-import Routes from '../../constants/routes'
 import { reapitConnectBrowserSession } from '../../core/connect-session'
 import { useReapitConnect, ReapitConnectSession } from '@reapit/connect-session'
 import { graphQLWrapper } from './__styles__/index'
 import { GRAPHQL_CONSTANTS } from '../../constants/graphql'
-import { openNewPage, ExternalPages, navigate } from '../../utils/navigation'
 
 const { scriptSrc, scriptId, localStorageKey, settings } = GRAPHQL_CONSTANTS
 
@@ -66,55 +48,19 @@ export const handleInitGQLPlayground =
   }
 
 export const GraphQLPage: FC = () => {
-  const location = useLocation()
-  const history = useHistory()
   const { connectSession } = useReapitConnect(reapitConnectBrowserSession)
   const [loaded, setLoaded] = useState(false)
   const graphQlRef = useRef<HTMLDivElement | null>(null)
-  const { pathname } = location
-
   useEffect(handleLoadGQLPlayground(setLoaded), [])
 
   useEffect(handleInitGQLPlayground(loaded, graphQlRef, connectSession), [graphQlRef, connectSession, loaded])
 
   return (
-    <FlexContainer isFlexAuto>
-      <SecondaryNavContainer>
-        <Title>API</Title>
-        <SecondaryNav className={elMb9}>
-          <SecondaryNavItem onClick={navigate(history, Routes.SWAGGER)} active={pathname === Routes.SWAGGER}>
-            REST API
-          </SecondaryNavItem>
-          <SecondaryNavItem onClick={navigate(history, Routes.WEBHOOKS_ABOUT)} active={pathname.includes('webhooks')}>
-            Webhooks
-          </SecondaryNavItem>
-          <SecondaryNavItem onClick={navigate(history, Routes.GRAPHQL)} active={pathname === Routes.GRAPHQL}>
-            GraphQL
-          </SecondaryNavItem>
-          <SecondaryNavItem onClick={navigate(history, Routes.DESKTOP)} active={pathname === Routes.DESKTOP}>
-            Desktop
-          </SecondaryNavItem>
-        </SecondaryNav>
-        <Icon className={elMb5} icon="graphQlInfographic" iconSize="large" />
-        <Subtitle>GraphQL Playground</Subtitle>
-        <SmallText hasGreyText>
-          GraphQL is a query language for APIs and a runtime for fulfilling those queries with your existing data.
-          GraphQL gives clients the power to ask for exactly what they need and nothing more, makes it easier to evolve
-          APIs over time, and enables powerful developer tools.
-        </SmallText>
-        <Button className={elMb5} intent="neutral" onClick={openNewPage(ExternalPages.graphQLDocs)}>
-          View Docs
-        </Button>
-        <Button className={elMb5} intent="critical" onClick={openNewPage(window.reapit.config.graphQLUri)}>
-          Open New
-        </Button>
-      </SecondaryNavContainer>
-      <PageContainer className={elHFull}>
-        <Title>GraphQL</Title>
-        {!loaded && <Loader label="Loading" fullPage />}
-        <div className={graphQLWrapper} ref={graphQlRef} />
-      </PageContainer>
-    </FlexContainer>
+    <>
+      <Title>GraphQL</Title>
+      {!loaded && <Loader label="Loading" fullPage />}
+      <div className={graphQLWrapper} ref={graphQlRef} />
+    </>
   )
 }
