--- packages/developer-portal/src/components/swagger/index.ts
@@ -1,2 +1,0 @@
-import Swagger from './swagger'
-export default Swagger
