+++ packages/developer-portal/src/components/webhooks/webhooks-manage.tsx
@@ -144,3 +144,5 @@ export const WebhooksManage: FC = () => {
     </>
   )
 }
+
+export default WebhooksManage
