--- packages/developer-portal/src/styles/elements/link.ts
@@ -1,15 +1,0 @@
-import { css } from '@linaria/core'
-import { bluePrimaryMedium } from '@/core/__styles__/colors'
-
-export const link = css`
-  color: ${bluePrimaryMedium};
-  font-weight: bold;
-`
-export const hyperlinked = css`
-  color: ${bluePrimaryMedium};
-  text-decoration: underline;
-  font-weight: bold;
-`
-export const linkNormal = css`
-  font-weight: normal;
-`
